fix: float to double

+++ ESGApiPlatform/Model/EsgApi.cs
@@ -8,24 +8,24 @@ namespace ESGApiPlatform.Model
     public class EsgApi
     {
 
-        public float? CurrentMonthDisbursement { get; set; }
-        public float? Last12MonthDisbursement { get; set; }
+        public double? CurrentMonthDisbursement { get; set; }
+        public double? Last12MonthDisbursement { get; set; }
         public int? Last12MonthBorrower { get; set; }
-        public float? Last12MonthDisbursementPercentageOnFemaleBorrower { get; set; }
-        public float? Last12MonthDisbursementPercentageOnMaleBorrower { get; set; }
-        public float? Last12MonthAvarageDisbursementOnFemaleBorrower { get; set; }
-        public float? Last12MonthAvarageDisbursementOnMaleBorrower { get; set; }
-        public float? YoYAvarageDisbursementOnFemaleBorrower { get; set; }
-        public float? YoYAvarageDisbursementOnMaleBorrower { get; set; }
-        public float? CurrentMonthDeposit { get; set; }
-        public float? Last12MonthDeposit { get; set; }
+        public double? Last12MonthDisbursementPercentageOnFemaleBorrower { get; set; }
+        public double? Last12MonthDisbursementPercentageOnMaleBorrower { get; set; }
+        public double? Last12MonthAvarageDisbursementOnFemaleBorrower { get; set; }
+        public double? Last12MonthAvarageDisbursementOnMaleBorrower { get; set; }
+        public double? YoYAvarageDisbursementOnFemaleBorrower { get; set; }
+        public double? YoYAvarageDisbursementOnMaleBorrower { get; set; }
+        public double? CurrentMonthDeposit { get; set; }
+        public double? Last12MonthDeposit { get; set; }
         public int? Last12MonthDepositor { get; set; }
-        public float? Last12MonthDepositPercentageOnFemaleDepositor { get; set; }
-        public float? Last12MonthDepositPercentageOnMaleDepositor { get; set; }
-        public float? Last12MonthAvarageDepositOnFemaleDepositor { get; set; }
-        public float? Last12MonthAvarageDepositOnMaleDepositor { get; set; }
-        public float? YoYAvarageDepositOnFemaleDepositor { get; set; }
-        public float? YoYAvarageLoanSizeOnMaleDepositor { get; set; }
+        public double? Last12MonthDepositPercentageOnFemaleDepositor { get; set; }
+        public double? Last12MonthDepositPercentageOnMaleDepositor { get; set; }
+        public double? Last12MonthAvarageDepositOnFemaleDepositor { get; set; }
+        public double? Last12MonthAvarageDepositOnMaleDepositor { get; set; }
+        public double? YoYAvarageDepositOnFemaleDepositor { get; set; }
+        public double? YoYAvarageLoanSizeOnMaleDepositor { get; set; }
         public DateTime? MonthStart { get; set; }
         public DateTime? MonthEnd { get; set; }
         public DateTime? YearStart { get; set; }
